fix: tighten pnpm install detection
