+++ src/resources/extensions/gsd/commands-handlers.ts
@@ -26,6 +26,7 @@ import { isAutoActive, checkRemoteAutoSession } from "./auto.js";
 import { getAutoWorktreePath } from "./auto-worktree.js";
 import { currentDirectoryRoot, projectRoot } from "./commands/context.js";
 import { loadPrompt } from "./prompt-loader.js";
+import { isPnpmInstall } from "../../shared/package-manager-detection.js";
 import {
   buildDoctorHealIssuePayload,
   buildDoctorHealSummary,
@@ -53,41 +54,6 @@ function isBunInstall(argv1: string | undefined = process.argv[1]): boolean {
   bunBinDirs.push(join(homedir(), ".bun", "bin"));
   const resolved = resolvePath(argv1);
   return bunBinDirs.some((dir) => resolved.startsWith(resolvePath(dir) + sep));
-}
-
-function hasPnpmPath(value: string | undefined): boolean {
-  if (!value) return false;
-  const normalized = value.replace(/\\/g, "/").toLowerCase();
-  return (
-    normalized.includes("/.pnpm/") ||
-    normalized.includes("/library/pnpm/") ||
-    normalized.endsWith("/pnpm") ||
-    normalized.endsWith("/pnpm.cjs") ||
-    normalized.endsWith("/pnpm.js")
-  );
-}
-
-function pathStartsWith(pathValue: string, dir: string): boolean {
-  const resolvedPath = resolvePath(pathValue);
-  const resolvedDir = resolvePath(dir);
-  return resolvedPath === resolvedDir || resolvedPath.startsWith(resolvedDir + sep);
-}
-
-function isPnpmInstall(
-  argv1: string | undefined = process.argv[1],
-  env: NodeJS.ProcessEnv = process.env,
-): boolean {
-  if (env.npm_config_user_agent?.startsWith("pnpm/")) return true;
-  if (hasPnpmPath(env.npm_execpath)) return true;
-  if (hasPnpmPath(argv1)) return true;
-  if (!argv1) return false;
-
-  const pnpmBinDirs: string[] = [];
-  if (env.PNPM_HOME) pnpmBinDirs.push(env.PNPM_HOME);
-  pnpmBinDirs.push(join(homedir(), "Library", "pnpm"));
-  pnpmBinDirs.push(join(homedir(), ".local", "share", "pnpm"));
-
-  return pnpmBinDirs.some((dir) => pathStartsWith(argv1, dir));
 }
 
 function resolveInstallCommand(pkg: string): string {
